requirements.txt
address==0.1.1
alembic==1.4.1
aliyun-python-sdk-core-v3==2.5.5
aliyun-python-sdk-dysmsapi==1.0.0
aliyun-python-sdk-push==3.10.0
aliyun-python-sdk-sts==3.0.0
aliyun-python-sdk-vod==2.9.0
amqp==2.5.2
article==0.1.1
attrs==19.2.0
Automat==0.8.0
backcall==0.1.0
beautifulsoup4==4.6.3
billiard==3.6.3.0
bleach==3.1.0
blinker==1.4
celery==4.3.0
certifi==2019.6.16
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
crcmod==1.7
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.1
django-allauth==0.39.1
django-ckeditor==5.4.0
django-js-asset==1.0.0
django-model-utils==3.1.2
django-mptt==0.10.0
django-notifications-hq==1.5.0
django-password-reset==2.0
django-redis==4.10.0
django-redis-cache==2.1.0
django-rest-auth==0.9.5
django-taggit==0.23.0
djangorestframework==3.11.0
dnspython==1.16.0
elastic-apm==4.0.2
elasticsearch==6.3.1
entrypoints==0.3
et-xmlfile==1.0.1
Flask==0.12.2
Flask-Login==0.5.0
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gevent==1.4.0
greenlet==0.4.15
grpcio==1.20.1
grpcio-tools==1.20.1
grpclib==0.3.1
gunicorn==19.7.1
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.6
importlib-metadata==0.23
incremental==17.5.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.5
json5==0.9.2
jsonfield==2.0.2
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
kiwisolver==1.1.0
kombu==4.6.3
lxml==4.2.5
Mako==1.1.2
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==3.2.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==7.2.0
multidict==4.7.5
mysqlclient==1.4.5
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.4
oauthlib==3.0.1
openpyxl==2.5.4
optionaldict==0.1.1
oss2==2.4.0
pandas==0.25.3
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.1
pickleshare==0.7.5
pika==0.13.0
pillow==5.3.0
Pillow-PIL==0.1.dev0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
Protego==0.1.16
protobuf==3.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pycryptodome==3.9.7
pycurl==7.43.0.5
PyDispatcher==2.0.5
pydub==0.23.1
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.7.1
PyMySQL==0.8.0
pyOpenSSL==19.1.0
pyparsing==2.4.5
pypinyin==0.37.0
pyquery==1.4.1
pyrsistent==0.15.4
pyspider==0.3.10
python-consul==1.1.0
python-dateutil==2.8.0
python-editor==1.0.4
python3-openid==3.1.0
pytz==2018.5
pywin32==225
pywinpty==0.5.5
PyYAML==5.2
pyzmq==18.1.0
qrcode==6.0
qtconsole==4.5.5
queuelib==1.5.0
redis==3.2.0
requests==2.18.4
requests-oauthlib==1.2.0
reverse==0.1.0
scipy==1.3.3
Scrapy==1.8.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
SQLAlchemy==1.2.2
sqlparse==0.3.0
tblib==1.6.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
tornado-sqlalchemy==0.6.1
traitlets==4.3.3
Twisted==19.10.0
u-msgpack-python==2.5.2
urllib3==1.22
utils==0.9.0
vine==1.3.0
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
wechatpy==1.6.1
Werkzeug==0.15.2
widgetsnbextension==3.5.1
wordcloud==1.6.0
WsgiDAV==3.0.2
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
xmltodict==0.12.0
zipp==0.6.0
zope.interface==4.7.1
